Create Python Turtle Beautiful Design
We will see one Python Turtle Beautiful Design which will be something like it will start drawing from the middle of the screen and will go towards the boundary of the screen using different colors. Python Turtle module is best for drawing and if we want our Kids to learn Programming with Python, then also Python Turtle is best as it is actually fun for kids as using Python Turtle anyone can create beautiful designs.